refactor: move restricted route to correct folder
